Quiet GitHub Actions console log output
This is really silly but Jepsen does not allow us to customize the console logger's appender levels
individually. Since we don't want to lose the jepsen.worker appender output (it's useful for seeing
the worker ops log with human-friendly timestamps, compared to the Jepsen time used in history.edn),
we just grep out the jepsen.worker output from stdout.